implement socket splicing on OpenBSD
socket splicing allows to do zero-copy data transfers between sockets.
This adds a specific implementation for OpenBSD using the setsockopt
SO_SPLICE and a default implementation with libevent (that is the old
code doing the copy in userland.)

It's possible to do the same on linux with splice(2), not implementd
though.